backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-docx>=1.1.0
python-pptx>=0.6.23
anthropic>=0.25.0
pydantic>=2.7.0
python-multipart>=0.0.9
openai>=1.0.0
docxtpl>=0.16.7
jinja2>=3.1.0
markdown>=3.6
beautifulsoup4>=4.12.0
